+++ TRLevelControl/IO/TRLevelWriter.cs
@@ -118,7 +118,7 @@ public class TRLevelWriter : BinaryWriter
 
     public void Write(TR1Entity entity)
     {
-        Write(entity.TypeID);
+        Write((short)entity.TypeID);
         Write(entity.Room);
         Write(entity.X);
         Write(entity.Y);
